services/data-retention-service/requirements.txt
psycopg2-binary==2.9.9
boto3==1.29.7



